api/requirements.txt
fastapi
uvicorn
jinja2
pysta
shapely
#gunicorn
#rasterio
#pyproj
